+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/s-thom/linkdrop/compare/v1.7.0...HEAD) - DATE
 
+### Fixed
+
+- Tags no longer appear out of order on links when searching
+
 ## [1.7.0](https://github.com/s-thom/linkdrop/compare/v1.6.3...v1.7.0) - 2023-04-09
 
 ### Added